add base class for text operations
